# openapi.yaml
openapi: 3.0.0

info:
  title: Ommatidia
  version: 0.3.0
  description: >-
    Unified and simplistic API to evaluate pupil detection algorithms for
    video-based eye tracking.
  contact:
    email: christopher@gundler.de
    name: Christopher Gundler
    url: https://gundler.de

paths:
  /:
    get:
      summary: Get the meta data of the detector
      responses:
        "200":
          description: Meta data regarding the pupil detection algorithm
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MetaData"

  /config/:
    get:
      summary: Define the configurations for the pupil detection algorithm as an JSON Schema.
      responses:
        "200":
          description: Configuration specified as an JSON Schema.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Config"

  /detections/:
    get:
      summary: Returns a list of running pupil detection algorithms.
      responses:
        "200":
          description: A list of "detectionId" for running pupil detection algorithms.
          content:
            application/json:
              schema:
                type: "array"
                items:
                  type: "integer"
      tags:
        - Detections
    post:
      summary: Initialize a new pupil detection algorithm with specific configuration.
      requestBody:
        $ref: "#/components/requestBodies/Config"
      responses:
        "200":
          description: Instance successfully created.
        "400":
          description: The given configuration is invalid.
          content:
            text/plain:
              schema:
                type: string

      tags:
        - Detections

  /detections/{detectionId}/:
    get:
      summary: Query an pupil detection algorithm and its configuration.
      parameters:
        - $ref: "#/components/parameters/DetectionId"
      responses:
        "200":
          description: The specified configuration on creation.
          content:
            application/json:
              schema:
                type: "object"
        "404":
          description: The selected ID was not found
          content:
            text/plain:
              schema:
                type: string
      tags:
        - Detections
    delete:
      summary: Delete the instance of the pupil detection algorithm and remove all associated resources.
      parameters:
        - $ref: "#/components/parameters/DetectionId"
      responses:
        "200":
          description: The detection was successfully deleted.
        "404":
          description: The selected ID was not found
      tags:
        - Detections
    post:
      summary: Evaluate a given image with the configured pupil detection algorihm.
      requestBody:
        $ref: "#/components/requestBodies/Sample"
      parameters:
        - $ref: "#/components/parameters/DetectionId"
      responses:
        "200":
          description: The estimated pupil.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Point"
                  - $ref: "#/components/schemas/Ellipse"
        "404":
          description: The selected ID was not found
      tags:
        - Detections

components:
  requestBodies:
    Sample:
      description: The sample encoded as an image.
      required: true
      content:
        image/png:
          schema:
            type: string
            format: binary
        image/jpg:
          schema:
            type: string
            format: binary
    Config:
      description: The configuration of the detection algorithm following the schema specified.
      required: true
      content:
        application/json:
          schema:
            type: "object"

  parameters:
    DetectionId:
      in: path
      name: detectionId
      schema:
        type: integer
        minimum: 0
      required: true
      description: Identifier for the running instance of pupil detection algorithm.

  schemas:
    Config:
      type: object
      description: "The definition of configuration regarding detector instances, i. e. the parameter. It must be a valid JSON schema."
      properties:
        type:
          type: "string"
          enum: ["object"]
          description: "Object definition as in https://json-schema.org/draft/2020-12/meta/core"
      additionalProperties: true
      required: ["type"]
    Confidence:
      type: number
      description: "The confidence of the algorithm in its prediction"
    Point:
      type: object
      properties:
        type:
          type: "string"
          enum: ["Point"]
        sample:
          $ref: "#/components/schemas/SampleDescription"
        x:
          $ref: "#/components/schemas/Position"
        y:
          $ref: "#/components/schemas/Position"
        confidence:
          $ref: "#/components/schemas/Confidence"
      required: [type, x, y, sample]
    Ellipse:
      type: object
      properties:
        type:
          type: "string"
          enum: ["Ellipse"]
        sample:
          $ref: "#/components/schemas/SampleDescription"
        x:
          $ref: "#/components/schemas/Position"
        y:
          $ref: "#/components/schemas/Position"
        major:
          type: number
          minimum: 0
        minor:
          type: number
          minimum: 0
        rotation:
          description: "Rotation in radians"
          type: number
          minimum: 0
          maximum: 6.28319
        confidence:
          $ref: "#/components/schemas/Confidence"
      required: [type, x, y, major, minor, rotation, sample]
    MetaData:
      type: object
      properties:
        name:
          type: string
          description: The name of the detector.
        additional_information:
          type: string
          description: The DOI of the publication.
        authors:
          type: array
          items:
            type: string
          description: The authors of the detector.
        license:
          type: string
          description: The license of the detector.
        prediction:
          type: string
          description: The output of the detector.
          enum:
            - Point
            - Ellipse
    Position:
      description: "Absolute position within the sample. The value must be smaller than the actual image dimension"
      type: number
      minimum: 0
    SampleDescription:
      type: object
      properties:
        width:
          type: number
          format: integer
          minimum: 0
        height:
          type: number
          format: integer
          minimum: 0

tags:
  - name: Detections
    description: Instances of the implemented pupil detection algorithm